+++ examples/data_warehouse/main.tf
@@ -16,10 +16,10 @@
 
 module "data_warehouse" {
   source  = "terraform-google-modules/bigquery/google//modules/data_warehouse"
-  version = "~> 8.0"
+  version = "~> 9.0"
 
   project_id          = var.project_id
-  region              = "asia-southeast1"
+  region              = "us-central1"
   deletion_protection = false
   force_destroy       = true
 }