+++ main.tf
@@ -29,16 +29,18 @@ locals {
 }
 
 resource "google_bigquery_dataset" "main" {
-  dataset_id                  = var.dataset_id
-  friendly_name               = var.dataset_name
-  description                 = var.description
-  location                    = var.location
-  delete_contents_on_destroy  = var.delete_contents_on_destroy
-  default_table_expiration_ms = var.default_table_expiration_ms
-  max_time_travel_hours       = var.max_time_travel_hours
-  project                     = var.project_id
-  labels                      = var.dataset_labels
-  resource_tags               = var.resource_tags
+  dataset_id                      = var.dataset_id
+  friendly_name                   = var.dataset_name
+  description                     = var.description
+  location                        = var.location
+  delete_contents_on_destroy      = var.delete_contents_on_destroy
+  default_table_expiration_ms     = var.default_table_expiration_ms
+  max_time_travel_hours           = var.max_time_travel_hours
+  storage_billing_model           = var.storage_billing_model
+  project                         = var.project_id
+  labels                          = var.dataset_labels
+  resource_tags                   = var.resource_tags
+  default_partition_expiration_ms = var.default_partition_expiration_ms
 
   dynamic "default_encryption_configuration" {
     for_each = var.encryption_key == null ? [] : [var.encryption_key]