+++ CHANGELOG.md
@@ -4,6 +4,57 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [10.1.1](https://github.com/terraform-google-modules/terraform-google-bigquery/compare/v10.1.0...v10.1.1) (2025-06-06)
+
+
+### Bug Fixes
+
+* update display metadata to fix regex for bigquery dataset id [#392](https://github.com/terraform-google-modules/terraform-google-bigquery/issues/392) ([2f7acb8](https://github.com/terraform-google-modules/terraform-google-bigquery/commit/2f7acb889f1439ecffba84b060f02d2392005558))
+* Upgrade gemin model used in data_warehouse ([#397](https://github.com/terraform-google-modules/terraform-google-bigquery/issues/397)) ([e383c94](https://github.com/terraform-google-modules/terraform-google-bigquery/commit/e383c942fb21da0d6592a306543558f79ae834f0))
+
+## [10.1.0](https://github.com/terraform-google-modules/terraform-google-bigquery/compare/v10.0.2...v10.1.0) (2025-03-26)
+
+
+### Features
+
+* Adding regex validation for big query dataset id and adding new output value computed_table_id ([#384](https://github.com/terraform-google-modules/terraform-google-bigquery/issues/384)) ([aab8238](https://github.com/terraform-google-modules/terraform-google-bigquery/commit/aab82382cbf67a3b7e1af801f80c3113592d4193))
+
+## [10.0.2](https://github.com/terraform-google-modules/terraform-google-bigquery/compare/v10.0.1...v10.0.2) (2025-03-14)
+
+
+### Bug Fixes
+
+* Use list(string) instead of tuple for some of the output variables ([#380](https://github.com/terraform-google-modules/terraform-google-bigquery/issues/380)) ([4d525f4](https://github.com/terraform-google-modules/terraform-google-bigquery/commit/4d525f453ac291c18f4843da4bbf1483daf4b72e))
+
+## [10.0.1](https://github.com/terraform-google-modules/terraform-google-bigquery/compare/v10.0.0...v10.0.1) (2025-03-06)
+
+
+### Bug Fixes
+
+* Make non-required object fields as optional for input variables ([#377](https://github.com/terraform-google-modules/terraform-google-bigquery/issues/377)) ([cd4ab5e](https://github.com/terraform-google-modules/terraform-google-bigquery/commit/cd4ab5e03d992556e7eba4860e716f742c142fb7))
+
+## [10.0.0](https://github.com/terraform-google-modules/terraform-google-bigquery/compare/v9.0.0...v10.0.0) (2025-03-03)
+
+
+### ⚠ BREAKING CHANGES
+
+* **deps:** Update Terraform terraform-google-modules/project-factory/google to v18 ([#366](https://github.com/terraform-google-modules/terraform-google-bigquery/issues/366))
+* **modules:** require TF v1.3+ ([#361](https://github.com/terraform-google-modules/terraform-google-bigquery/issues/361))
+* **TPG>=6.11:** add deletion protection variable to google_workflows_workflow ([#370](https://github.com/terraform-google-modules/terraform-google-bigquery/issues/370))
+
+### Features
+
+* Add output variable env_vars ([#375](https://github.com/terraform-google-modules/terraform-google-bigquery/issues/375)) ([ab9ed8c](https://github.com/terraform-google-modules/terraform-google-bigquery/commit/ab9ed8c95fcb8db9517c1fc22aa52de7371bdfe8))
+* add storage_billing_model and default_partition_expiration_ms ([#367](https://github.com/terraform-google-modules/terraform-google-bigquery/issues/367)) ([53646de](https://github.com/terraform-google-modules/terraform-google-bigquery/commit/53646dedc8167c17b9b609d2ed2d5ef37668f7bb))
+
+
+### Bug Fixes
+
+* **deps:** Update Terraform provider constraints ([#357](https://github.com/terraform-google-modules/terraform-google-bigquery/issues/357)) ([c285f34](https://github.com/terraform-google-modules/terraform-google-bigquery/commit/c285f3478a1cd65af1b0812515c2bb3484e9fa3b))
+* **deps:** Update Terraform terraform-google-modules/project-factory/google to v18 ([#366](https://github.com/terraform-google-modules/terraform-google-bigquery/issues/366)) ([d7723c4](https://github.com/terraform-google-modules/terraform-google-bigquery/commit/d7723c495d6bf735e883543a2e83703877107368))
+* **modules:** require TF v1.3+ ([#361](https://github.com/terraform-google-modules/terraform-google-bigquery/issues/361)) ([9f9581b](https://github.com/terraform-google-modules/terraform-google-bigquery/commit/9f9581b1fb066f938d2af611d6ece1fa162381fe))
+* **TPG>=6.11:** add deletion protection variable to google_workflows_workflow ([#370](https://github.com/terraform-google-modules/terraform-google-bigquery/issues/370)) ([7233527](https://github.com/terraform-google-modules/terraform-google-bigquery/commit/72335270b153d959f0c3f46932f2970aaf5acdcc))
+
 ## [9.0.0](https://github.com/terraform-google-modules/terraform-google-bigquery/compare/v8.1.0...v9.0.0) (2024-10-16)